+++ examples/common/database.go
@@ -8,15 +8,15 @@ import (
 
 // MutableTable is a table which its entities are mutable.
 type MutableTable struct {
-	UID          int64     `db:"uid"           json:"uid"`           // UID is the unique id.
-	CT           time.Time `db:"ct"            json:"ct"`            // CT is the create time.
-	ModifiedTime null.Time `db:"modified_time" json:"modified_time"` // ModifiedTime is the modified time.
-	DT           int64     `db:"dt"            json:"dt"`            // DT is the delete time.
-	V            int64     `db:"v"             json:"v"`             // V is the version.
+	UID       int64     `db:"uid"        json:"uid"`        // UID is the unique id.
+	CreatedAt time.Time `db:"created_at" json:"created_at"` // CreatedAt is the create time.
+	UpdatedAt null.Time `db:"updated_at" json:"updated_at"` // UpdatedAt is the update time.
+	DeletedAt int64     `db:"deleted_at" json:"deleted_at"` // DeletedAt is the delete time.
+	Version   int64     `db:"version"    json:"version"`    // Version is the row version.
 }
 
 // ImmutableTable is a table which its entities are immutable.
 type ImmutableTable struct {
-	ID int64     `db:"id" json:"id"` // ID is the primary key.
-	CT null.Time `db:"ct" json:"ct"` // CT is the create time.
+	ID        int64     `db:"id"         json:"id"`         // ID is the primary key.
+	CreatedAt null.Time `db:"created_at" json:"created_at"` // CreatedAt is the create time.
 }