+++ examples/database/org.go
@@ -5,8 +5,8 @@ import "github.com/tmoeish/tsq/examples/common"
 // 组织表
 // @TABLE(
 //
-//	ux=[{fields=["Name"]}]
-//	ct
+//	ux=[{fields=["Name"]}],
+//	created_at
 //
 // )
 type Org struct {